requirements.txt
Django==2.2.7
djangorestframework==3.10.3
gunicorn==20.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==3.3.1
